requirements.txt
numpy>=1.19.0
opencv-python>=4.2.0.34
pyzmq>=19.0.1
vidgear>=0.1.8
